# processes/order/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8091",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/rest/task/update": {
      "post": {
        "tags": [
          "complete-task-controller"
        ],
        "operationId": "updateTask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTaskDtoObject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskIdDto"
                }
              }
            }
          }
        }
      }
    },
    "/rest/task/load": {
      "post": {
        "tags": [
          "load-task-controller"
        ],
        "operationId": "loadData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTaskDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/HtmlFormDto"
                    },
                    {
                      "$ref": "#/components/schemas/JsonFormDto"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/rest/task/complete": {
      "post": {
        "tags": [
          "complete-task-controller"
        ],
        "operationId": "completeTask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTaskDtoObject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskIdDto"
                }
              }
            }
          }
        }
      }
    },
    "/rest/process/start": {
      "post": {
        "tags": [
          "start-process-controller"
        ],
        "operationId": "placeOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceOrderDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        }
      }
    },
    "/rest/process/start/form": {
      "get": {
        "tags": [
          "start-process-controller"
        ],
        "operationId": "loadItems",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/HtmlFormDto"
                    },
                    {
                      "$ref": "#/components/schemas/JsonFormDto"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CompleteTaskDtoObject": {
        "required": [
          "formData",
          "userTask"
        ],
        "type": "object",
        "properties": {
          "userTask": {
            "$ref": "#/components/schemas/UserTaskDto"
          },
          "formData": {
            "type": "object"
          }
        }
      },
      "UserTaskDto": {
        "required": [
          "bpmnProcessId",
          "elementId",
          "key",
          "processDefinitionKey",
          "processInstanceKey",
          "taskState",
          "variables"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "integer",
            "format": "int64"
          },
          "elementId": {
            "type": "string"
          },
          "processInstanceKey": {
            "type": "integer",
            "format": "int64"
          },
          "bpmnProcessId": {
            "type": "string"
          },
          "processDefinitionKey": {
            "type": "integer",
            "format": "int64"
          },
          "variables": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "taskState": {
            "type": "string"
          },
          "assignee": {
            "type": "string"
          }
        }
      },
      "TaskIdDto": {
        "required": [
          "taskId"
        ],
        "type": "object",
        "properties": {
          "taskId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "FormDtoObject": {
        "required": [
          "updatable"
        ],
        "type": "object",
        "properties": {
          "formData": {
            "type": "object"
          },
          "updatable": {
            "type": "boolean"
          }
        }
      },
      "HtmlFormDto": {
        "required": [
          "html",
          "updatable"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/FormDtoObject"
          },
          {
            "type": "object",
            "properties": {
              "html": {
                "type": "string"
              }
            }
          }
        ]
      },
      "JsonFormDto": {
        "required": [
          "schema",
          "uiSchema",
          "updatable"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/FormDtoObject"
          },
          {
            "type": "object",
            "properties": {
              "schema": {
                "type": "string"
              },
              "uiSchema": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ItemDto": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "image": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PlaceOrderDto": {
        "required": [
          "city",
          "email",
          "firstname",
          "items",
          "lastname",
          "street",
          "zip"
        ],
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "zip": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemDto"
            }
          }
        }
      },
      "MessageDto": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}